+++ tinnsleep/create_reports.py
@@ -6,9 +6,9 @@ from tinnsleep.classification import AmplitudeThresholding
 from tinnsleep.check_impedance import create_annotation_mne, Impedance_thresholding_sliding, check_RMS, \
     fuse_with_classif_result
 from tinnsleep.signal import rms
-from tinnsleep.scoring import generate_bruxism_report
 from tinnsleep.signal import is_good_epochs
 from scipy.interpolate import interp1d
+from tinnsleep.scoring import generate_bruxism_report, create_list_events, classif_to_burst, burst_to_episode
 
 
 def preprocess(raw, picks_chan, picks_imp, duration, interval, params, THR_imp=6000, get_log=False, filter="default"):
@@ -179,3 +179,82 @@ def merge_labels_list(list_valid_labels, n_epoch_final, merge_fun=np.all):
             valid_labels[:, i] = f1(xnew)
 
     return merge_fun(valid_labels, axis=-1)
+
+def _cond_subclassif(ep_to_sub, labels_artifacts, labels_condition, time_interval):
+    # ---------------Subclassifying MEMA and bruxism events----------------------
+    comb_ep = []
+    pure_ep = []
+    compt_arti = 0
+    # merge episodes and compare with condition and artefacts
+    for elm in ep_to_sub:
+        if np.sum(labels_artifacts[int(elm.beg / time_interval):int(elm.end / time_interval)]) == \
+                (int(elm.end / time_interval) - int(elm.beg / time_interval)):
+            print("no artifacts detected")
+            if np.sum(labels_condition[int(elm.beg / time_interval):int(elm.end / time_interval)]) > 0:
+                comb_ep.append(elm)
+            else:
+                pure_ep.append(elm)
+        else:
+            compt_arti += 1
+    # ------------------
+    # Pure episodes creation
+    li_ep_p = create_list_events(pure_ep, time_interval, time_interval * len(labels_condition), boolean_output=True)
+    # Combined episode creation
+    li_ep_c = create_list_events(comb_ep, time_interval, time_interval * len(labels_condition), boolean_output=True)
+
+    return li_ep_c, li_ep_p, compt_arti
+
+def _labels_to_ep_and_bursts(labels, time_interval, delim_ep, min_burst_joining=3):
+    # ------------grouping episodes and bursts together------------------------------------
+    bursts = classif_to_burst(labels, time_interval=time_interval)
+    li_ep = burst_to_episode(bursts, delim=delim_ep, min_burst_joining=min_burst_joining)
+    events = create_list_events(li_ep, time_interval, len(labels) * time_interval, boolean_output=True)
+    ep_and_bursts = np.any(np.c_[labels, events], axis=-1)  # rassembling bruxism bursts and episodes
+    return ep_and_bursts, li_ep
+
+
+
+
+def combine_brux_MEMA(labels_brux, labels_artifacts_brux, time_interval_brux, delim_ep_brux, labels_MEMA,
+                      labels_artifacts_MEMA, time_interval_MEMA, delim_ep_MEMA,
+                      min_burst_joining_brux=3, min_burst_joining_MEMA=0):
+    """
+    Hypothesis labels_artifacts_brux has the same length as labels_brux idem for MEMA
+
+    """
+    # Putting labels inputs on the same sampling and epoching
+    if len(labels_brux) != len(labels_MEMA):
+        if len(labels_brux) < len(labels_MEMA):
+            labels_brux = merge_labels_list ([labels_brux], len(labels_MEMA))
+            # adapts and fuses artifacts
+            labels_artifacts =  merge_labels_list ([labels_artifacts_brux,labels_artifacts_MEMA], len(labels_MEMA))
+            time_interval = time_interval_MEMA
+        else:
+            labels_MEMA = merge_labels_list([labels_MEMA], len(labels_brux))
+            # adapts and fuses artifacts
+            labels_artifacts = merge_labels_list([labels_artifacts_brux,labels_artifacts_MEMA], len(labels_brux))
+            time_interval = time_interval_brux
+    else: #inputs of same length
+        # fuses artifacts
+        labels_artifacts = merge_labels_list([labels_artifacts_brux, labels_artifacts_MEMA], len(labels_brux))
+        time_interval = time_interval_brux
+
+    # Creating lists of episode and bursts for bruxism and MEMA
+    brux_burst_ep, li_ep_brux = _labels_to_ep_and_bursts(labels_brux, time_interval, delim_ep_brux,
+                                                         min_burst_joining=min_burst_joining_brux)
+    MEMA_burst_ep, li_ep_MEMA = _labels_to_ep_and_bursts(labels_MEMA, time_interval, delim_ep_MEMA,
+                                                         min_burst_joining=min_burst_joining_MEMA)
+
+    # Conditionnal labelling of events
+    MEMA_comb_ep, MEMA_pure_ep, compt_arti_MEMA= _cond_subclassif(li_ep_MEMA, labels_artifacts,
+                                                                 brux_burst_ep, time_interval)
+    brux_comb_ep, brux_pure_ep, compt_arti_brux = _cond_subclassif(li_ep_brux, labels_artifacts,
+                                                                   MEMA_burst_ep, time_interval)
+
+    return brux_comb_ep, brux_pure_ep, compt_arti_brux, MEMA_comb_ep, MEMA_pure_ep, compt_arti_MEMA
+
+
+
+
+
+
